trait CreateConnection
