+++ crates/resolver/src/name_server/name_server.rs
@@ -7,7 +7,7 @@
 
 use std::cmp::Ordering;
 use std::fmt::{self, Debug, Formatter};
-use std::net::SocketAddr;
+use std::future::Future;
 use std::pin::Pin;
 use std::sync::Arc;
 use std::time::Instant;
@@ -17,39 +17,47 @@ use futures_util::stream::{once, Stream};
 
 #[cfg(feature = "mdns")]
 use proto::multicast::MDNS_IPV4;
-use proto::op::NoopMessageFinalizer;
-use proto::tcp::TcpClientStream;
-use proto::udp::UdpClientStream;
-use proto::xfer::{DnsExchange, DnsHandle, DnsRequest, DnsResponse, FirstAnswer};
-use proto::DnsMultiplexer;
+use proto::xfer::{DnsHandle, DnsRequest, DnsResponse, FirstAnswer};
 use tracing::debug;
 
-use crate::config::Protocol;
 use crate::config::{NameServerConfig, ResolverOpts};
 use crate::error::ResolveError;
-use crate::name_server::connection_provider::{ConnectionConnect, ConnectionFuture};
 use crate::name_server::{GenericConnection, NameServerState, NameServerStats, RuntimeProvider};
 #[cfg(feature = "mdns")]
 use proto::multicast::{MdnsClientConnect, MdnsClientStream, MdnsQueryType};
 
-/// Specifies the details of a remote NameServer used for lookups
+/// This struct is needed only for testing. Specifically, `C` is needed for mocking.
 #[derive(Clone)]
-pub struct NameServer<P: RuntimeProvider + Send> {
+pub struct AbstractNameServer<
+    C: DnsHandle<Error = ResolveError> + Send + Sync + 'static + CreateConnection,
+    P: RuntimeProvider + Send,
+> {
     config: NameServerConfig,
     options: ResolverOpts,
-    client: Arc<Mutex<Option<GenericConnection>>>,
+    client: Arc<Mutex<Option<C>>>,
     state: Arc<NameServerState>,
     stats: Arc<NameServerStats>,
     runtime_provider: P,
 }
 
-impl<P: RuntimeProvider> Debug for NameServer<P> {
+/// Specifies the details of a remote NameServer used for lookups
+pub type NameServer<P> = AbstractNameServer<GenericConnection, P>;
+
+impl<C, P> Debug for AbstractNameServer<C, P>
+where
+    C: DnsHandle<Error = ResolveError> + Send + Sync + 'static + CreateConnection,
+    P: RuntimeProvider + Send,
+{
     fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), fmt::Error> {
         write!(f, "config: {:?}, options: {:?}", self.config, self.options)
     }
 }
 
-impl<P: RuntimeProvider> NameServer<P> {
+impl<C, P> AbstractNameServer<C, P>
+where
+    C: DnsHandle<Error = ResolveError> + Send + Sync + 'static + CreateConnection,
+    P: RuntimeProvider + Send,
+{
     /// Construct a new Nameserver with the configuration and options. The connection provider will create UDP and TCP sockets
     pub fn new(config: NameServerConfig, options: ResolverOpts, runtime_provider: P) -> Self {
         Self {
@@ -66,7 +74,7 @@ impl<P: RuntimeProvider> NameServer<P> {
     pub fn from_conn(
         config: NameServerConfig,
         options: ResolverOpts,
-        client: GenericConnection,
+        client: C,
         runtime_provider: P,
     ) -> Self {
         Self {
@@ -94,7 +102,7 @@ impl<P: RuntimeProvider> NameServer<P> {
     /// This will return a mutable client to allows for sending messages.
     ///
     /// If the connection is in a failed state, then this will establish a new connection
-    async fn connected_mut_client(&mut self) -> Result<GenericConnection, ResolveError> {
+    async fn connected_mut_client(&mut self) -> Result<C, ResolveError> {
         let mut client = self.client.lock().await;
 
         // if this is in a failure state
@@ -104,7 +112,8 @@ impl<P: RuntimeProvider> NameServer<P> {
             // TODO: we need the local EDNS options
             self.state.reinit(None);
 
-            let new_client = self.new_connection(&self.config, &self.options).await?;
+            let new_client =
+                C::new_connection(&self.runtime_provider, &self.config, &self.options).await?;
 
             // establish a new connection
             *client = Some(new_client);
@@ -115,139 +124,6 @@ impl<P: RuntimeProvider> NameServer<P> {
         Ok((*client)
             .clone()
             .expect("bad state, client should be connected"))
-    }
-
-    fn new_connection(
-        &self,
-        config: &NameServerConfig,
-        options: &ResolverOpts,
-    ) -> ConnectionFuture<P> {
-        let dns_connect = match config.protocol {
-            Protocol::Udp => {
-                let provider_handle = self.runtime_provider.clone();
-                let closure = move |addr: SocketAddr| provider_handle.bind_udp(addr);
-                let stream = UdpClientStream::with_creator(
-                    config.socket_addr,
-                    None,
-                    options.timeout,
-                    Arc::new(closure),
-                );
-                let exchange = DnsExchange::connect(stream);
-                ConnectionConnect::Udp(exchange)
-            }
-            Protocol::Tcp => {
-                let socket_addr = config.socket_addr;
-                let timeout = options.timeout;
-                let tcp_future = self.runtime_provider.connect_tcp(socket_addr);
-
-                let (stream, handle) =
-                    TcpClientStream::with_future(tcp_future, socket_addr, timeout);
-                // TODO: need config for Signer...
-                let dns_conn = DnsMultiplexer::with_timeout(
-                    stream,
-                    handle,
-                    timeout,
-                    NoopMessageFinalizer::new(),
-                );
-
-                let exchange = DnsExchange::connect(dns_conn);
-                ConnectionConnect::Tcp(exchange)
-            }
-            #[cfg(feature = "dns-over-tls")]
-            Protocol::Tls => {
-                let socket_addr = config.socket_addr;
-                let timeout = options.timeout;
-                let tls_dns_name = config.tls_dns_name.clone().unwrap_or_default();
-                let tcp_future = self.runtime_provider.connect_tcp(socket_addr);
-
-                #[cfg(feature = "dns-over-rustls")]
-                let client_config = config.tls_config.clone();
-
-                #[cfg(feature = "dns-over-rustls")]
-                let (stream, handle) = {
-                    crate::tls::new_tls_stream_with_future(
-                        tcp_future,
-                        socket_addr,
-                        tls_dns_name,
-                        client_config,
-                    )
-                };
-                #[cfg(not(feature = "dns-over-rustls"))]
-                let (stream, handle) =
-                    { crate::tls::new_tls_stream::<R>(socket_addr, bind_addr, tls_dns_name) };
-
-                let dns_conn = DnsMultiplexer::with_timeout(
-                    stream,
-                    handle,
-                    timeout,
-                    NoopMessageFinalizer::new(),
-                );
-
-                let exchange = DnsExchange::connect(dns_conn);
-                ConnectionConnect::Tls(exchange)
-            }
-            #[cfg(feature = "dns-over-https")]
-            Protocol::Https => {
-                let socket_addr = config.socket_addr;
-                let tls_dns_name = config.tls_dns_name.clone().unwrap_or_default();
-                #[cfg(feature = "dns-over-rustls")]
-                let client_config = config.tls_config.clone();
-                let tcp_future = self.runtime_provider.connect_tcp(socket_addr);
-
-                let exchange = crate::https::new_https_stream_with_future(
-                    tcp_future,
-                    socket_addr,
-                    tls_dns_name,
-                    client_config,
-                );
-                ConnectionConnect::Https(exchange)
-            }
-            #[cfg(feature = "dns-over-quic")]
-            Protocol::Quic => {
-                let socket_addr = config.socket_addr;
-                let bind_addr = config.bind_addr.unwrap_or(match socket_addr {
-                    SocketAddr::V4(_) => SocketAddr::new(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)), 0),
-                    SocketAddr::V6(_) => {
-                        SocketAddr::new(IpAddr::V6(Ipv6Addr::new(0, 0, 0, 0, 0, 0, 0, 0)), 0)
-                    }
-                });
-                let tls_dns_name = config.tls_dns_name.clone().unwrap_or_default();
-                #[cfg(feature = "dns-over-rustls")]
-                let client_config = config.tls_config.clone();
-                let udp_future = self.runtime_provider.bind_udp(bind_addr);
-
-                let exchange = crate::quic::new_quic_stream_with_future(
-                    udp_future,
-                    socket_addr,
-                    tls_dns_name,
-                    client_config,
-                );
-                ConnectionConnect::Quic(exchange)
-            }
-            #[cfg(feature = "mdns")]
-            Protocol::Mdns => {
-                let socket_addr = config.socket_addr;
-                let timeout = options.timeout;
-
-                let (stream, handle) =
-                    MdnsClientStream::new(socket_addr, MdnsQueryType::OneShot, None, None, None);
-                // TODO: need config for Signer...
-                let dns_conn = DnsMultiplexer::with_timeout(
-                    stream,
-                    handle,
-                    timeout,
-                    NoopMessageFinalizer::new(),
-                );
-
-                let exchange = DnsExchange::connect(dns_conn);
-                ConnectionConnect::Mdns(exchange)
-            }
-        };
-
-        ConnectionFuture {
-            connect: dns_connect,
-            spawner: self.runtime_provider.create_handle(),
-        }
     }
 
     async fn inner_send<R: Into<DnsRequest> + Unpin + Send + 'static>(
@@ -297,8 +173,9 @@ impl<P: RuntimeProvider> NameServer<P> {
     }
 }
 
-impl<P> DnsHandle for NameServer<P>
+impl<C, P> DnsHandle for AbstractNameServer<C, P>
 where
+    C: DnsHandle<Error = ResolveError> + Send + Sync + 'static + CreateConnection,
     P: RuntimeProvider,
 {
     type Response = Pin<Box<dyn Stream<Item = Result<DnsResponse, ResolveError>> + Send>>;
@@ -316,7 +193,11 @@ where
     }
 }
 
-impl<P: RuntimeProvider> Ord for NameServer<P> {
+impl<C, P> Ord for AbstractNameServer<C, P>
+where
+    C: DnsHandle<Error = ResolveError> + Send + Sync + 'static + CreateConnection,
+    P: RuntimeProvider + Send,
+{
     /// Custom implementation of Ord for NameServer which incorporates the performance of the connection into it's ranking
     fn cmp(&self, other: &Self) -> Ordering {
         // if they are literally equal, just return
@@ -328,20 +209,33 @@ impl<P: RuntimeProvider> Ord for NameServer<P> {
     }
 }
 
-impl<P: RuntimeProvider> PartialOrd for NameServer<P> {
+impl<C, P> PartialOrd for AbstractNameServer<C, P>
+where
+    C: DnsHandle<Error = ResolveError> + Send + Sync + 'static + CreateConnection,
+    P: RuntimeProvider + Send,
+{
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         Some(self.cmp(other))
     }
 }
 
-impl<P: RuntimeProvider> PartialEq for NameServer<P> {
+impl<C, P> PartialEq for AbstractNameServer<C, P>
+where
+    C: DnsHandle<Error = ResolveError> + Send + Sync + 'static + CreateConnection,
+    P: RuntimeProvider + Send,
+{
     /// NameServers are equal if the config (connection information) are equal
     fn eq(&self, other: &Self) -> bool {
         self.config == other.config
     }
 }
 
-impl<P: RuntimeProvider> Eq for NameServer<P> {}
+impl<C, P> Eq for AbstractNameServer<C, P>
+where
+    C: DnsHandle<Error = ResolveError> + Send + Sync + 'static + CreateConnection,
+    P: RuntimeProvider + Send,
+{
+}
 
 // TODO: once IPv6 is better understood, also make this a binary keep.
 #[cfg(feature = "mdns")]
@@ -363,6 +257,19 @@ where
         bind_addr: None,
     };
     NameServer::new_with_provider(config, options, conn_provider)
+}
+
+/// Used for connection mocking
+pub trait CreateConnection: Sized {
+    type FutureConn<P: RuntimeProvider>: Future<Output = Result<Self, ResolveError>>
+        + Send
+        + 'static;
+
+    fn new_connection<P: RuntimeProvider>(
+        runtime_provider: &P,
+        config: &NameServerConfig,
+        options: &ResolverOpts,
+    ) -> Self::FutureConn<P>;
 }
 
 #[cfg(test)]
